+++ tests/data/test_import_from_affymetrix.py
@@ -191,10 +191,14 @@ class TestImportFromAffyReport(AffyTestMixin, unittest.TestCase):
     def setUp(self):
         self.runner = CliRunner()
         self.report = DATA_DIR / "affyreport.txt"
+        self.report_no_cols = DATA_DIR / "affyreport_nocols.txt"
 
-    def link_files(self, working_dir):
+    def link_files(self, working_dir, report=None):
         report = working_dir / "affyreport.txt"
+        report_no_cols = working_dir / "affyreport_nocols.txt"
+
         report.symlink_to(self.report)
+        report_no_cols.symlink_to(self.report_no_cols)
 
     def test_help(self):
         result = self.runner.invoke(import_from_affymetrix, ["--help"])
@@ -319,6 +323,66 @@ class TestImportFromAffyReport(AffyTestMixin, unittest.TestCase):
             locus_list = plink_file.get_loci()
 
             self.assertEqual(len(sample_list), 1)
+            self.assertEqual(len(locus_list), 1)
+
+    @patch('src.features.smarterdb.Dataset.result_dir',
+           new_callable=PropertyMock)
+    @patch('src.features.smarterdb.Dataset.working_dir',
+           new_callable=PropertyMock)
+    def test_import_from_affymetrix_no_check(
+            self, my_working_dir, my_result_dir):
+        # create a temporary directory using the context manager
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            working_dir = pathlib.Path(tmpdirname)
+            results_dir = working_dir / "results"
+
+            # assign return value to mocked property
+            my_working_dir.return_value = working_dir
+            my_result_dir.return_value = results_dir
+
+            # copy test data files
+            self.link_files(working_dir)
+
+            result = self.runner.invoke(
+                import_from_affymetrix,
+                [
+                    "--dataset",
+                    "test.zip",
+                    "--report",
+                    "affyreport_nocols.txt",
+                    "--chip_name",
+                    self.chip_name,
+                    "--assembly",
+                    "OAR3",
+                    "--breed_code",
+                    "TEX",
+                    "--sample_field",
+                    "alias",
+                    "--create_samples",
+                    "--src_version",
+                    "Oar_v4.0",
+                    "--src_imported_from",
+                    "affymetrix",
+                    "--coding",
+                    "ab",
+                    "--skip_coordinate_check"
+                ]
+            )
+
+            self.assertEqual(0, result.exit_code, msg=result.exception)
+            self.assertEqual(SampleSheep.objects.count(), 2)
+
+            # check imported chip_name attribute
+            for sample in SampleSheep.objects:
+                self.assertEqual(sample.chip_name, self.chip_name)
+
+            plink_path = results_dir / "OAR3" / "affyreport_nocols_updated"
+            plink_file = plinkfile.open(str(plink_path))
+
+            sample_list = plink_file.get_samples()
+            locus_list = plink_file.get_loci()
+
+            self.assertEqual(len(sample_list), 2)
             self.assertEqual(len(locus_list), 1)